requirements.txt.txt
Flask==3.0.3
qrcode==7.4.2
pillow==10.4.0
python-dotenv==1.0.1
gunicorn==23.0.0
twilio==9.3.2
